Extract common mock chart data into separate constant
