+++ client/web/src/enterprise/insights/components/creation-ui-kit/live-preview/index.ts
@@ -1,2 +1,3 @@
 export * from './LivePreviewCard'
+export * from './constants'
 export { useLivePreview, StateStatus } from './use-live-preview'
