+++ client/web/src/enterprise/insights/pages/insights/creation/capture-group/components/CaptureGroupCreationContent.tsx
@@ -8,7 +8,7 @@ import { useAsyncInsightTitleValidator } from '../../../../../components/form/ho
 import { useField } from '../../../../../components/form/hooks/useField'
 import { FormChangeEvent, SubmissionErrors, useForm } from '../../../../../components/form/hooks/useForm'
 import { createRequiredValidator } from '../../../../../components/form/validators'
-import { Insight } from '../../../../../core/types'
+import { Insight } from '../../../../../core'
 import {
     repositoriesExistValidator,
     repositoriesFieldValidator,
@@ -18,7 +18,7 @@ import { CaptureGroupFormFields } from '../types'
 import { searchQueryValidator } from '../utils/search-query-validator'
 
 import { CaptureGroupCreationForm } from './CaptureGoupCreationForm'
-import { CaptureGroupCreationLivePreview } from './live-preview/CaptureGroupCreationLivePreview'
+import { CaptureGroupCreationLivePreview } from './CaptureGroupCreationLivePreview'
 
 const INITIAL_VALUES: CaptureGroupFormFields = {
     repositories: '',
