--- client/web/src/enterprise/insights/pages/insights/creation/search-insight/components/live-preview-chart/index.ts
@@ -1,2 +1,0 @@
-export { SearchInsightLivePreview } from './SearchInsightLivePreview'
-export { DEFAULT_MOCK_CHART_CONTENT } from './constant'
